+++ xd/xdfs/xlsxiterator.cpp
@@ -114,7 +114,7 @@ void XlsxIterator::skip(int delta)
 
     if ((unsigned int)m_current_row > m_file->getRowCount())
     {
-        m_current_row = m_file->getRowCount();
+        m_current_row = (int)m_file->getRowCount();
         m_eof = true;
         return;
     }
@@ -134,7 +134,7 @@ void XlsxIterator::goLast()
 {
     m_bof = false;
     m_eof = false;
-    m_current_row = m_file->getRowCount();
+    m_current_row = (int)m_file->getRowCount();
     goRow(m_current_row);
 }
 
@@ -343,7 +343,7 @@ bool XlsxIterator::modifyStructure(const xd::StructureModify& mod_params, xd::IJ
             dai->type = it->params.type;
             dai->width = it->params.width;
             dai->scale = it->params.scale;
-            dai->ordinal = m_fields.size();
+            dai->ordinal = (int)m_fields.size();
             dai->expr_text = it->params.expression;
             dai->expr = parse(it->params.expression);
             m_fields.push_back(dai);
@@ -368,7 +368,7 @@ bool XlsxIterator::modifyStructure(const xd::StructureModify& mod_params, xd::IJ
             dai->type = it->params.type;
             dai->width = it->params.width;
             dai->scale = it->params.scale;
-            dai->ordinal = m_fields.size();
+            dai->ordinal = (int)m_fields.size();
             dai->expr_text = it->params.expression;
             dai->expr = parse(it->params.expression);
             m_fields.insert(m_fields.begin()+insert_idx, dai);
